app/services/scraper.py
import requests
from bs4 import BeautifulSoup
from app.utils.logger import logger, AuditLog

class CompanyScraperScraper:
    @staticmethod
    def search_buyers_trujillo() -> list:
        """
        Busca empresas compradoras de minerales en Trujillo, Perú.
        Consulta directorios y portales locales respetando robots.txt.
        """
        buyers = []
        
        # Fuentes sugeridas (placeholder - adaptar a fuentes reales)
        sources = [
            {
                "name": "Cámara de Comercio La Libertad",
                "url": "https://www.camaralibertad.org.pe",
                "type": "chamber"
            },
            {
                "name": "MINEM - Registro de Concesiones",
                "url": "https://portal.minem.gob.pe",
                "type": "government"
            }
        ]
        
        for source in sources:
            try:
                headers = {"User-Agent": "Mozilla/5.0 (Linux; Android 10)"}
                response = requests.get(source["url"], headers=headers, timeout=10)
                
                AuditLog.log_api_call(source["name"], source["url"], response.status_code)
                
                if response.status_code == 200:
                    soup = BeautifulSoup(response.content, "html.parser")
                    # Parsear según estructura del sitio (adaptar CSS selectors)
                    # Placeholder: retorna estructura vacía
                    logger.info(f"Scraped {source['name']}")
            except Exception as e:
                logger.error(f"Error scraping {source['name']}: {str(e)}")
                AuditLog.log_api_call(source["name"], source["url"], error=str(e))
        
        return buyers
