requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML
google-generativeai==0.3.0

# Task Queue (optional for production)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
